+++ worker_health/fitness-awsmetal-all.sh
@@ -3,4 +3,4 @@
 set -e
 # set -x
 
-./fitness.py -p gecko-t -hh -s t-linux-metal $@
+./fitness_check.py -p gecko-t -hh -s t-linux-metal "$@"
