+++ worker_health/fitness-gecko-t-win10-64-1803-hw.sh
@@ -3,6 +3,6 @@
 set -e
 # set -x
 
-./fitness.py \
+./fitness_check.py \
   -p releng-hardware \
   gecko-t-win10-64-1803-hw "$@"
