+++ worker_health/worker_health/quarantine.py
@@ -3,8 +3,9 @@
 import json
 import os
 
-import fitness
 import taskcluster
+
+from worker_health import fitness
 
 # see https://github.com/mozilla-platform-ops/relops-infra/blob/master/quarantine_tc.py
 # for prior art
