+++ worker_health/worker_health/tc_helpers.py
@@ -1,6 +1,7 @@
-import utils
+from worker_health import utils
 
 
+# extracted functions from Fitness
 class TCHelper:
     def __init__(
         self,
